refactor: cleanup md parser logic
