+++ packages/rich-text-from-markdown/package.json
@@ -33,13 +33,13 @@
   "dependencies": {
     "@contentful/rich-text-types": "^12.1.2",
     "remark-parse": "^6.0.1",
-    "unified": "^7.0.0"
+    "unified": "^7.0.0",
+    "lodash": "^4.17.11"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.119",
     "faker": "^4.1.0",
     "jest": "^23.6.0",
-    "lodash": "^4.17.11",
     "rimraf": "^2.6.2",
     "rollup": "^0.66.6",
     "rollup-plugin-commonjs": "^9.2.0",
