+++ packages/rich-text-from-markdown/src/types/index.ts
@@ -1,24 +1,14 @@
-declare module 'unified' {
-  interface MarkdownNode {
-    depth: string;
-    type: string;
-    ordered: Boolean;
-    children: MarkdownNode[];
-    content: MarkdownNode[];
-    value: string;
-  }
-
-  interface MarkdownTree {
-    children: MarkdownNode[];
-  }
-
-  interface Processor {
-    parse(md: string): MarkdownTree;
-  }
-
-  export default function(): { use(md: Object, config: Object): Processor };
+export interface MarkdownNode extends MarkdownTree {
+  depth: string;
+  type: string;
+  ordered: Boolean;
+  value: string;
 }
 
-declare module 'remark-parse' {
-  export default function(one: Object, two: Object): void;
+export interface MarkdownTree {
+  children: MarkdownNode[];
+}
+
+export interface MarkdownLinkNode extends MarkdownNode {
+  url: string;
 }
